// web/src/features/dashboard/types.ts
export interface QuotaDataItem {
  id?: number
  user_id?: number
  username?: string
  display_name?: string
  avatar_url?: string
  model_name?: string
  created_at: number
  token_used?: number
  uncached_input_tokens?: number
  uncached_output_tokens?: number
  cache_read_tokens?: number
  cache_write_tokens?: number
  count?: number
  quota?: number
}
// ...
export interface DashboardFilters {
  start_timestamp?: Date
  end_timestamp?: Date
  time_granularity?: TimeGranularity
  username?: string
}